Responding to @isaknaslundbh comments
