+++ Robot_Adapter/CRUD/Create/Properties/SectionProperties.cs
