+++ Robot_Adapter/CRUD/Read/Elements/Bars.cs
