+++ Robot_Adapter/CRUD/Read/Elements/Nodes.cs
