+++ Robot_Adapter/CRUD/Read/Elements/Openings.cs
