+++ Robot_Adapter/CRUD/Read/Elements/Panels.cs
