+++ Robot_Adapter/CRUD/Read/Loads/Loads.cs
