+++ Robot_Adapter/CRUD/Read/Properties/Labels.cs
