+++ Robot_Adapter/CRUD/Read/Properties/LinkConstraints.cs
