+++ Robot_Adapter/CRUD/Read/Properties/Releases.cs
