+++ Robot_Adapter/CRUD/Read/Properties/SectionProperties.cs
