+++ Robot_Adapter/Robot_Adapter.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Convert\FromRobot\Geometry\Polyline.cs" />
     <Compile Include="Convert\FromRobot\Geometry\Point.cs" />
     <Compile Include="Convert\FromRobot\Geometry\Polycurve.cs" />
+    <Compile Include="Convert\ToRobot\Loads\Load.cs" />
     <Compile Include="Convert\FromRobot\Properties\LinearRelease.cs" />
     <Compile Include="Convert\FromRobot\Properties\BarRelease.cs" />
     <Compile Include="Convert\FromRobot\Properties\Material.cs" />
@@ -156,12 +157,12 @@
     <Compile Include="Convert\ToRobot\Geometry\Curve.cs" />
     <Compile Include="Convert\ToRobot\Groups and Lists\Group.cs" />
     <Compile Include="Convert\ToRobot\Groups and Lists\Lists.cs" />
-    <Compile Include="Convert\ToRobot\Loads\Load.cs" />
+    <Compile Include="Convert\FromRobot\Loads\Load.cs" />
     <Compile Include="Convert\ToRobot\Properties\Material.cs" />
     <Compile Include="Convert\ToRobot\Properties\MemberType.cs" />
-    <Compile Include="Convert\ToRobot\Elements\Panel.cs" />
+    <Compile Include="Convert\FromRobot\Elements\Panel.cs" />
     <Compile Include="Convert\ToRobot\Properties\LinearRelease.cs" />
-    <Compile Include="Convert\ToRobot\Properties\Property2D.cs" />
+    <Compile Include="Convert\ToRobot\Properties\SurfaceProperty.cs" />
     <Compile Include="Convert\ToRobot\Properties\SectionProperty_Concrete.cs" />
     <Compile Include="Convert\ToRobot\Properties\SectionProperty_Steel.cs" />
     <Compile Include="Convert\ToRobot\Properties\Support.cs" />
@@ -169,27 +170,27 @@
     <Compile Include="Convert\ToRobot\Properties\SectionProperty.cs" />
     <Compile Include="CRUD\Create\Properties\LinearReleases.cs" />
     <Compile Include="CRUD\Create\Properties\Supports.cs" />
-    <Compile Include="CRUD\Groups and Lists\DesignGroup.cs" />
+    <Compile Include="CRUD\Groups and Lists\DesignGroups.cs" />
     <Compile Include="CRUD\Create\Properties\FramingElementDesignProperties.cs" />
-    <Compile Include="CRUD\Create\Properties\LinkConstraint.cs" />
+    <Compile Include="CRUD\Create\Properties\LinkConstraints.cs" />
     <Compile Include="CRUD\Create\Elements\Bars.cs" />
     <Compile Include="CRUD\Create\Elements\FEMeshes.cs" />
     <Compile Include="CRUD\Create\Properties\BarReleases.cs" />
     <Compile Include="CRUD\Create\Loads\LoadCombination.cs" />
-    <Compile Include="CRUD\Create\Properties\Property2D.cs" />
+    <Compile Include="CRUD\Create\Properties\SurfaceProperties.cs" />
     <Compile Include="CRUD\Groups and Lists\Groups.cs" />
     <Compile Include="CRUD\Create\Elements\Panels.cs" />
     <Compile Include="CRUD\Create\Loads\Loads.cs" />
     <Compile Include="CRUD\Create\Elements\Nodes.cs" />
     <Compile Include="CRUD\Create\Loads\Loadcase.cs" />
-    <Compile Include="CRUD\Create\Properties\Material.cs" />
-    <Compile Include="CRUD\Create\Properties\SectionProperty.cs" />
+    <Compile Include="CRUD\Create\Properties\Materials.cs" />
+    <Compile Include="CRUD\Create\Properties\SectionProperties.cs" />
     <Compile Include="CRUD\Create\Properties\RigidLinks.cs" />
     <Compile Include="CRUD\Delete\_Delete.cs" />
     <Compile Include="CRUD\Create\_Create.cs" />
     <Compile Include="AdapterActions\Execute.cs" />
-    <Compile Include="CRUD\Read\Loads\LoadCombination.cs" />
-    <Compile Include="CRUD\Read\Properties\Label.cs" />
+    <Compile Include="CRUD\Read\Loads\LoadCombinations.cs" />
+    <Compile Include="CRUD\Read\Properties\Labels.cs" />
     <Compile Include="CRUD\Read\Properties\LinearReleases.cs" />
     <Compile Include="CRUD\Update\Elements\Nodes.cs" />
     <Compile Include="CRUD\Update\Loads\LoadCombinations.cs" />
@@ -200,23 +201,23 @@
     <Compile Include="CRUD\Update\Properties\LinearReleases.cs" />
     <Compile Include="CRUD\Update\Elements\Panels.cs" />
     <Compile Include="CRUD\Update\Properties\Materials.cs" />
-    <Compile Include="CRUD\Update\Properties\Properties2D.cs" />
+    <Compile Include="CRUD\Update\Properties\SurfaceProperties.cs" />
     <Compile Include="Types\NextId.cs" />
-    <Compile Include="CRUD\Read\Elements\Bar.cs" />
+    <Compile Include="CRUD\Read\Elements\Bars.cs" />
     <Compile Include="CRUD\Read\Elements\Groups.cs" />
-    <Compile Include="CRUD\Read\Elements\Mesh.cs" />
-    <Compile Include="CRUD\Read\Elements\RigidLink.cs" />
-    <Compile Include="CRUD\Read\Loads\Load.cs" />
-    <Compile Include="CRUD\Read\Elements\Node.cs" />
+    <Compile Include="CRUD\Read\Elements\Meshes.cs" />
+    <Compile Include="CRUD\Read\Elements\RigidLinks.cs" />
+    <Compile Include="CRUD\Read\Loads\Loads.cs" />
+    <Compile Include="CRUD\Read\Elements\Nodes.cs" />
     <Compile Include="CRUD\Read\Properties\Supports.cs" />
-    <Compile Include="CRUD\Read\Elements\Opening.cs" />
-    <Compile Include="CRUD\Read\Elements\Panel.cs" />
-    <Compile Include="CRUD\Read\Properties\LinkConstraint.cs" />
-    <Compile Include="CRUD\Read\Properties\Property2D.cs" />
-    <Compile Include="CRUD\Read\Properties\SectionProperty.cs" />
-    <Compile Include="CRUD\Read\Properties\Material.cs" />
-    <Compile Include="CRUD\Read\Properties\Release.cs" />
-    <Compile Include="CRUD\Read\Loads\Loadcase.cs" />
+    <Compile Include="CRUD\Read\Elements\Openings.cs" />
+    <Compile Include="CRUD\Read\Elements\Panels.cs" />
+    <Compile Include="CRUD\Read\Properties\LinkConstraints.cs" />
+    <Compile Include="CRUD\Read\Properties\SurfaceProperties.cs" />
+    <Compile Include="CRUD\Read\Properties\SectionProperties.cs" />
+    <Compile Include="CRUD\Read\Properties\Materials.cs" />
+    <Compile Include="CRUD\Read\Properties\Releases.cs" />
+    <Compile Include="CRUD\Read\Loads\Loadcases.cs" />
     <Compile Include="CRUD\Read\Elements\DesignGroups.cs" />
     <Compile Include="CRUD\Read\Properties\FramingElementDesignProperties.cs" />
     <Compile Include="CRUD\Read\Results\BarResults.cs" />
